% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSDnsANCOVA.R
\name{LSDnsANCOVA}
\alias{LSDnsANCOVA}
\title{Neutrosophic Analysis of Covariance for Latin Square Design}
\usage{
LSDnsANCOVA(Lower_y, Upper_y, Lower_z, Upper_z, design, alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{Lower_y}{Matrix containing lower bounds of response observations.}

\item{Upper_y}{Matrix containing upper bounds of response observations.}

\item{Lower_z}{Matrix containing lower bounds of covariate observations.}

\item{Upper_z}{Matrix containing upper bounds of covariate observations.}

\item{design}{Matrix representing Latin square treatment allocation.}

\item{alpha}{Significance level for the F-test and LSD test.Default is 0.05.}

\item{verbose}{Logical. If TRUE, displays the ANCOVA table,
LSD interval, treatment comparisons and significance codes.
Default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{nancova_table}: Neutrosophic ANCOVA table.
\item \code{comparison}: LSD treatment comparisons, if applicable.
\item \code{LSD}: Lower and upper limits of the LSD interval.
}
}
\description{
Performs Neutrosophic Analysis of Covariance (NANCOVA) for Latin
square designs using interval-valued response and covariate
observations. The function computes neutrosophic sums of squares,
mean squares, interval-valued F-statistics, significance tests,
and Least Significant Difference (LSD)-based treatment
comparisons. For crisp data, enter identical lower and upper
values to obtain the corresponding classical ANCOVA results.
}
\details{
Input matrix structure:
\itemize{
  \item Rows represent blocks (or rows of the design).
  \item Columns represent treatment positions within each block.
  \item `Lower_y` and `Upper_y` must have the same dimensions as
  the design matrix.
  \item `Lower_z` and `Upper_z` must have the same dimensions as
  the design matrix.
}
}
\examples{
Lower_y <- matrix(c(
18.86,15.49,18.62,16.37,12.04,
18.23,26.80,26.70,15.89,16.96,
26.26,12.16,27.15,18.30,22.93,
25.31,26.71,18.36,9.98,28.34,
29.78,21.20,21.99,18.97,18.71
), nrow = 5, byrow = TRUE)

Upper_y <- matrix(c(
21.14,22.51,25.38,21.63,19.96,
27.77,29.20,33.30,18.11,25.04,
29.74,19.84,36.85,25.70,31.07,
34.69,29.29,21.64,18.02,33.66,
32.22,28.80,24.01,25.03,21.29
), nrow = 5, byrow = TRUE)

Lower_z <- matrix(c(
9.11,7.39,7.09,4.10,5.92,
6.91,4.26,13.10,5.86,12.54,
8.89,4.98,4.32,8.80,9.77,
6.40,7.28,3.29,3.81,4.45,
6.42,2.51,6.06,3.25,5.73
), nrow = 5, byrow = TRUE)

Upper_z <- matrix(c(
16.89,10.61,14.91,9.90,8.08,
11.09,9.74,16.90,12.14,17.46,
13.11,11.02,11.68,11.20,18.23,
9.60,12.72,10.71,8.19,7.55,
15.58,7.49,9.94,12.75,14.27
), nrow = 5, byrow = TRUE)

design <- matrix(c(
2,1,5,4,3,
1,4,3,5,2,
3,2,4,1,5,
5,3,1,2,4,
4,5,2,3,1
), nrow = 5, byrow = TRUE)

LSDnsANCOVA(Lower_y, Upper_y, Lower_z, Upper_z, design, alpha = 0.05, verbose = TRUE)

}
