% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSDnsANOVA.R
\name{LSDnsANOVA}
\alias{LSDnsANOVA}
\title{Neutrosophic Analysis of Variance for Latin Square Design}
\usage{
LSDnsANOVA(Lower_y, Upper_y, design, alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{Lower_y}{Matrix containing lower bounds of observations.}

\item{Upper_y}{Matrix containing upper bounds of observations.}

\item{design}{Matrix representing Latin square treatment allocation.}

\item{alpha}{Significance level (default is 0.05.) for the F-test and LSD test.}

\item{verbose}{Logical. If TRUE, displays the NANOVA table,
LSD interval, treatment comparisons, and significance codes.
Default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{nanova_table}: Neutrosophic ANOVA table.
\item \code{comparison}: LSD treatment comparisons, if applicable.
\item \code{LSD}: Lower and upper limits of the LSD interval.
}
}
\description{
Performs Neutrosophic Analysis of Variance (NANOVA) for Latin square
designs using interval-valued observations. The function computes
neutrosophic sums of squares, mean squares, interval-valued F-statistics,
significance tests, and Least Significant Difference (LSD)-based
treatment comparisons. When the data are crisp, identical lower and
upper values may be entered to obtain the corresponding classical
ANOVA results.
}
\details{
Input matrix structure:
\itemize{
  \item Rows represent blocks (or rows of the design).
  \item Columns represent treatment positions within each block.
  \item `Lower_y` and `Upper_y` must have the same dimensions as
  the design matrix.
}
}
\examples{
Lower_y <- matrix(c(
243.28,233.45,232.41,239.27,
237.38,223.30,240.86,240.33,
232.12,234.67,230.45,220.81,
219.26,231.43,248.44,227.88
), nrow = 4, byrow = TRUE)

Upper_y <- matrix(c(
246.72,238.55,237.59,242.73,
242.62,226.70,245.14,245.67,
235.88,239.33,233.55,223.19,
224.74,236.57,253.57,230.12
), nrow = 4, byrow = TRUE)

design <- matrix(c(
3,4,2,1,
1,2,4,3,
4,3,1,2,
2,1,3,4
), nrow = 4, byrow = TRUE)

LSDnsANOVA(Lower_y, Upper_y, design, alpha = 0.05, verbose = TRUE)

}
