% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YSDnsANCOVA.R
\name{YSDnsANCOVA}
\alias{YSDnsANCOVA}
\title{Neutrosophic Analysis of Covariance for Youden Square Design}
\usage{
YSDnsANCOVA(Lower_y, Upper_y, Lower_z, Upper_z, design, alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{Lower_y}{Matrix containing lower bounds of response observations.}

\item{Upper_y}{Matrix containing upper bounds of response observations.}

\item{Lower_z}{Matrix containing lower bounds of covariate observations.}

\item{Upper_z}{Matrix containing upper bounds of covariate observations.}

\item{design}{Matrix representing Youden square treatment allocation.}

\item{alpha}{Significance level for the F-test and LSD test. Default is 0.05.}

\item{verbose}{Logical. If TRUE, prints the ANCOVA table,
LSD interval, treatment comparisons and significance codes.
Default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{nancova_table}: Neutrosophic ANCOVA table.
\item \code{comparison}: LSD treatment comparisons, if performed.
\item \code{LSD}: Lower and upper limits of the LSD interval.
}
}
\description{
Performs Neutrosophic Analysis of Covariance (NANCOVA) for Youden
square designs using interval-valued observations and covariates.
The function computes neutrosophic sums of squares, mean squares,
interval-valued F-statistics, significance tests, and Least
Significant Difference (LSD)-based treatment comparisons.
For crisp data, enter identical lower and upper values to obtain
the corresponding classical ANCOVA results.
}
\details{
Input matrix structure:
\itemize{
  \item Rows represent blocks of the design.
  \item Columns represent treatment positions within each row.
  \item `Lower_y` and `Upper_y` must have the same dimensions as
  the design matrix.
  \item `Lower_z` and `Upper_z` must have the same dimensions as
  the design matrix.
}
}
\examples{
Lower_y <- matrix(c(
2.06,2.34,0.13,15.63,15.72,8.18,292.48,
15.39,218.67,10.02,13.67,8.90,25.23,24.86,
35.78,24.90,308.36,19.96,22.01,20.17,28.99,
45.05,42.44,33.19,280.58,31.40,31.82,28.41
), nrow = 4, byrow = TRUE)

Upper_y <- matrix(c(
5.94,8.26,2.33,18.17,18.08,12.42,295.52,
19.61,221.33,14.38,17.33,13.10,27.77,29.54,
38.22,27.10,311.64,25.44,26.39,22.63,33.60,
48.55,45.96,35.41,283.42,36.00,35.58,32.59
), nrow = 4, byrow = TRUE)

Lower_z <- matrix(c(
10.57,269.93,224.08,260.05,257.81,257.58,13.76,
246.72,7.42,216.00,257.15,232.38,237.91,215.32,
246.50,215.17,2.63,257.26,267.90,212.10,254.15,
262.67,250.51,250.45,6.09,245.01,228.31,258.90
), nrow = 4, byrow = TRUE)

Upper_z <- matrix(c(
13.43,272.07,227.92,265.95,260.19,260.42,18.24,
249.28,12.58,220.00,262.85,235.62,240.09,218.68,
251.50,220.83,7.37,260.74,272.10,215.90,257.85,
265.33,253.49,253.55,11.91,250.99,231.69,263.10
), nrow = 4, byrow = TRUE)

design <- matrix(c(
2,3,4,5,6,7,1,
7,1,2,3,4,5,6,
6,7,1,2,3,4,5,
5,6,7,1,2,3,4
), nrow = 4, byrow = TRUE)

YSDnsANCOVA(Lower_y, Upper_y, Lower_z, Upper_z, design, alpha = 0.05,  verbose = TRUE)

}
