% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YSDnsANOVA.R
\name{YSDnsANOVA}
\alias{YSDnsANOVA}
\title{Neutrosophic Analysis of Variance for Youden Square Design}
\usage{
YSDnsANOVA(Lower_y, Upper_y, design, alpha = 0.05,  verbose = FALSE)
}
\arguments{
\item{Lower_y}{Matrix containing lower bounds of observations.}

\item{Upper_y}{Matrix containing upper bounds of observations.}

\item{design}{Matrix representing Youden square treatment allocation.}

\item{alpha}{Significance level for the F-test and LSD test. Default is 0.05.}

\item{verbose}{Logical. If TRUE, displays the NANOVA table,
LSD interval, treatment comparisons and significance codes.
Default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{nanova_table}: Neutrosophic ANOVA table.
\item \code{comparison}: LSD treatment comparisons, if applicable.
\item \code{LSD}: Lower and upper limits of the LSD interval.
}
}
\description{
Performs Neutrosophic Analysis of Variance (NANOVA) for Youden
square designs using interval-valued observations. The function
computes neutrosophic sums of squares, mean squares,
interval-valued F-statistics, significance tests, and Least
Significant Difference (LSD)-based treatment comparisons.
For crisp data, enter identical lower and upper values to obtain
the corresponding classical ANOVA results.
}
\details{
Input matrix structure:
\itemize{
  \item Rows represent blocks of the design.
  \item Columns represent treatment positions within each row.
  \item `Lower_y` and `Upper_y` must have the same dimensions as
  the design matrix.
}
}
\examples{
Lower_y <- matrix(c(
4.51,7.77,3.53,7.92,5.97,7.42,6.37,
6.37,4.36,4.68,8.19,6.08,6.80,4.86,
4.88,4.64,8.69,7.81,9.50,9.60,9.10,
8.05,7.92,5.61,8.41,6.53,11.36,5.94
), nrow = 4, byrow = TRUE)

Upper_y <- matrix(c(
8.12,11.06,6.44,11.16,8.30,10.17,9.81,
10.31,7.50,8.01,10.82,8.49,10.69,7.77,
7.76,7.24,10.82,11.04,13.20,11.63,11.59,
11.66,10.45,8.92,12.03,8.95,14.64,9.14
), nrow = 4, byrow = TRUE)

design <- matrix(c(
2,3,4,5,6,7,1,
7,1,2,3,4,5,6,
6,7,1,2,3,4,5,
5,6,7,1,2,3,4
), nrow = 4, byrow = TRUE)

YSDnsANOVA(Lower_y, Upper_y, design, alpha = 0.05,  verbose = TRUE)

}
